+++ dflib/src/test/java/com/nhl/dflib/SeriesGroupByTest.java
@@ -1,5 +1,6 @@
 package com.nhl.dflib;
 
+import com.nhl.dflib.unit.DFAsserts;
 import com.nhl.dflib.unit.SeriesAsserts;
 import org.junit.Test;
 
@@ -11,5 +12,47 @@ public class SeriesGroupByTest {
                 .group((String s) -> s.length());
 
         new SeriesAsserts(gb.toSeries()).expectData("a", "b", "e", "cd", "fg");
+    }
+
+    @Test
+    public void testAgg() {
+
+        Series<String> aggregated = Series.forData("a", "b", "cd", "e", "fg")
+                .group((String s) -> s.length())
+                .agg(SeriesAggregator.concat("_"));
+
+        new SeriesAsserts(aggregated).expectData("a_b_e", "cd_fg");
+    }
+
+    @Test
+    public void testAggMultiple() {
+
+        DataFrame aggregated = Series.forData("a", "b", "cd", "e", "fg")
+                .group((String s) -> s.length())
+                .aggMultiple(
+                        SeriesAggregator.first(),
+                        SeriesAggregator.concat("|"),
+                        SeriesAggregator.concat("_"));
+
+        new DFAsserts(aggregated, "first", "concat", "concat_")
+                .expectHeight(2)
+                .expectRow(0, "a", "a|b|e", "a_b_e")
+                .expectRow(1, "cd", "cd|fg", "cd_fg");
+    }
+
+    @Test
+    public void testAggMultiple_Named() {
+
+        DataFrame aggregated = Series.forData("a", "b", "cd", "e", "fg")
+                .group((String s) -> s.length())
+                .aggMultiple(
+                        SeriesAggregator.first().named("f"),
+                        SeriesAggregator.concat("|").named("c1"),
+                        SeriesAggregator.concat("_").named("c2"));
+
+        new DFAsserts(aggregated, "f", "c1", "c2")
+                .expectHeight(2)
+                .expectRow(0, "a", "a|b|e", "a_b_e")
+                .expectRow(1, "cd", "cd|fg", "cd_fg");
     }
 }