+++ dflib/src/test/java/com/nhl/dflib/Series_AggTest.java
@@ -1,0 +1,43 @@
+package com.nhl.dflib;
+
+import com.nhl.dflib.unit.SeriesAsserts;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Collection;
+
+import static org.junit.Assert.*;
+
+@RunWith(Parameterized.class)
+public class Series_AggTest extends BaseObjectSeriesTest {
+
+    public Series_AggTest(BaseObjectSeriesTest.SeriesTypes seriesType) {
+        super(seriesType);
+    }
+
+    @Parameterized.Parameters
+    public static Collection<Object[]> data() {
+        return ALL_SERIES_TYPES;
+    }
+
+    @Test
+    public void testAgg() {
+
+        String aggregated = createSeries("a", "b", "cd", "e", "fg").agg(SeriesAggregator.concat("_"));
+
+        assertEquals("a_b_cd_e_fg", aggregated);
+    }
+
+    @Test
+    public void testAggMultiple() {
+
+        Series<?> aggregated = createSeries("a", "b", "cd", "e", "fg")
+                .aggMultiple(
+                        SeriesAggregator.first(),
+                        SeriesAggregator.concat("|"),
+                        SeriesAggregator.concat("_"));
+
+        new SeriesAsserts(aggregated).expectData("a", "a|b|cd|e|fg", "a_b_cd_e_fg");
+    }
+}