+++ RELEASE-NOTES.md
@@ -38,6 +38,7 @@
 * #60 dflib-jdbc: JdbcConnector to support external transaction management
 * #62 Series.filter(ValuePredicate), Series.filter(BooleanSeries) 
 * #63 Basic `Series.groupBy(..)` 
+* #64 SeriesGroupBy aggregation functionality 
 
 ## Release 0.5